Fix ver service selector
The kustomize builtin 'commonLabels' changes the service selectors in
addition to the resource labels, which means that the service does not
match any pods when they are deployed at run-time. This is because the
automatically deployed pods do not have the labels that were added by
'commonLabels'. Use a transformer instead to just add labels to metadata.